+++ src/mol-gl/webgl/context.ts
@@ -289,7 +289,7 @@ export interface WebGLContext {
     /** Cache for textures, managed by consumers */
     readonly namedTextures: { [name: string]: Texture }
 
-    createRenderTarget: (width: number, height: number, depth?: boolean, type?: 'uint8' | 'float32' | 'fp16', filter?: TextureFilter, format?: 'rgba' | 'alpha', depthStencil?: boolean) => RenderTarget
+    createRenderTarget: (width: number, height: number, depthStencil?: 'none' | 'depth' | 'depth-stencil', type?: 'uint8' | 'float32' | 'fp16', filter?: TextureFilter, format?: 'rgba' | 'alpha') => RenderTarget
     createDrawTarget: () => RenderTarget
     bindDrawingBuffer: () => void
     getDrawingBufferSize: () => { width: number, height: number }
@@ -441,8 +441,8 @@ export function createContext(gl: GLRenderingContext, props: Partial<{ pixelScal
             pixelScale = value;
         },
 
-        createRenderTarget: (width: number, height: number, depth?: boolean, type?: 'uint8' | 'float32' | 'fp16', filter?: TextureFilter, format?: 'rgba' | 'alpha', depthStencil?: boolean) => {
-            const renderTarget = createRenderTarget(gl, resources, width, height, depth, type, filter, format, depthStencil);
+        createRenderTarget: (width: number, height: number, depthStencil?: 'none' | 'depth' | 'depth-stencil', type?: 'uint8' | 'float32' | 'fp16', filter?: TextureFilter, format?: 'rgba' | 'alpha') => {
+            const renderTarget = createRenderTarget(gl, resources, width, height, depthStencil, type, filter, format);
             renderTargets.add(renderTarget);
             return {
                 ...renderTarget,